pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "a public api from ShanTou.University"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "pydantic-settings>=2.3.1",
    "toml>=0.10.2",
    "python-barcode>=0.15.1",
    "pillow>=11.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"
[[project.authors]]
name = "ShanTou-U"
email = "dev@ShanTou.University"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
unit = "scripts:mk_unit"
rotate = "scripts:mk_rotate"
init = "scripts:init_project_details"
start = "api.__main__:main"

[tool.rye]
managed = true
dev-dependencies = [ "coverage>=7.5.4", "pytest>=8.2.2",]

[tool.ruff]
fix = true
line-length = 80
include = [ "pyproject.toml", "src/**/*.py", "scripts/**/*.py",]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [ "E", "F", "UP", "B", "SIM", "I",]

[tool.rye.scripts.dev]
cmd = "python -m api"

[tool.rye.scripts.dev.env]
RELOAD = "True"
LOG_LEVEL = "debug"

[tool.hatch.build.targets.wheel]
packages = [ "src/api", "scripts",]
